# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quartermaster API",
    "description": "Self-hosted kitchen inventory management.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/auth/login": {
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_logout",
        "responses": {
          "204": {
            "description": ""
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/me": {
      "get": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPair"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/auth/switch-household": {
      "post": {
        "tags": [
          "accounts"
        ],
        "operationId": "auth_switch_household",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SwitchHouseholdRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable error code, machine-readable."
          },
          "message": {
            "type": "string",
            "description": "Human-readable description. Not localised; clients should prefer `code`."
          },
          "unrestorable_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Present on `code = batch_not_restorable` when the failure came from a\nbulk call — lists the specific batches the caller needs to look at.\nAbsent on single-batch restore (the path already names the batch)."
          }
        }
      },
...
      "HouseholdDto": {
        "type": "object",
        "required": [
          "id",
          "name",
          "timezone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "device_label": {
            "type": [
              "string",
              "null"
            ]
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "MeHouseholdDto": {
        "type": "object",
        "required": [
          "household",
          "role",
          "joined_at"
        ],
        "properties": {
          "household": {
            "$ref": "#/components/schemas/HouseholdDto"
          },
          "joined_at": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/MembershipRole"
          }
        }
      },
      "MeResponse": {
        "type": "object",
        "required": [
          "user",
          "households"
        ],
        "properties": {
          "household_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "household_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "household_timezone": {
            "type": [
              "string",
              "null"
            ]
          },
          "households": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MeHouseholdDto"
            }
          },
          "public_base_url": {
            "type": [
              "string",
              "null"
            ]
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
...
      "MembershipRole": {
        "type": "string",
        "enum": [
          "admin",
          "member"
        ]
      },
...
      "RefreshRequest": {
        "type": "object",
        "required": [
          "refresh_token"
        ],
        "properties": {
          "refresh_token": {
            "type": "string"
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "device_label": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional label applied to the refresh token (shown on `/auth/me`)."
          },
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "invite_code": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required unless the server is in `first_run_only` mode and no users\nexist yet, or in `open` mode."
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "SwitchHouseholdRequest": {
        "type": "object",
        "required": [
          "household_id"
        ],
        "properties": {
          "household_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "TokenPair": {
        "type": "object",
        "required": [
          "access_token",
          "refresh_token",
          "token_type",
          "expires_in"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "integer",
            "format": "int64"
          },
          "refresh_token": {
            "type": "string"
          },
          "token_type": {
            "type": "string"
          }
        }
      },
...
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "properties": {
          "email": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    }
  },
...
}